Facilate session driven display of interactions in assistant app
I altered assistant.html and assistant/views.py. Created middleware.py in familyhack5 folder, which sets a variable named first_interaction_id in the session as soon as a new session is started. This variable is set to -1, which signifies that there have been no interactions yet. As soon as the user posts the first question, the first_interaction_id is set to the value of the id of the newly created interaction model.